*Linux
In this image, the primary objective was to set up a MySQL database instance within a Docker container. This deployment was a crucial part of the overall project infrastructure. I created a dedicated Docker network named "wp-mysql-network" to facilitate communication between containers, launched a Docker container named "mysql-db" using the mysql:5.6 image, attaching it to the custom network. The MySQL container was successfully deployed and connected to the specified network.